semantic-learning/pyproject.toml
[project]
name = "semantic-learning"
version = "0.1.0"
description = ""
authors = [
    {name = "David Moore",email = "damoo@microsoft.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "semantic-kernel (>=1.18.2,<2.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "azure-identity (>=1.19.0,<2.0.0)",
    "azure-search (>=1.0.0b2,<2.0.0)",
    "azure-search-documents (>=11.5.2,<12.0.0)",
    "ollama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
